# docs/specs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stratum — organizations, accounts, billing, and notifications.",
        "title": "Stratum API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/organizations/{organizationID}/audit-log": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Page-based (limit/page) unless a cursor is supplied, in which case it switches to cursor pagination. Filterable by actor/action/resource/from/to. Admin/owner only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "List organization audit log entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "max results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number, ignored once cursor is set (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pagination cursor from a previous response; switches to cursor mode",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by actor",
                        "name": "actor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by resource",
                        "name": "resource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 lower bound on created_at",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 upper bound on created_at",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/EventRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/organizations/{organizationID}/audit-log/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Streams the organization's audit log as a CSV attachment, filterable by actor/action/resource/from/to. Admin/owner only.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Export organization audit log as CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "filter by actor",
                        "name": "actor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by resource",
                        "name": "resource",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 lower bound on created_at",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 upper bound on created_at",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "EventRecord": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "actor": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "organization_id": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
...
        "Payload": {
            "type": "object",
            "properties": {
                "data": {},
                "next_cursor": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/pagination"
                },
                "status": {
                    "$ref": "#/definitions/Status"
                }
            }
        },
...
        "Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token issued by Supabase auth.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}